Fix two bugs: Update time in interactive mode, fix watch style.
http://developer.android.com/training/wearables/watch-faces/drawing.html
claims that "The system calls the Engine.onTimeTick() method every
minute."  This is only true in ambient mode according to the
WatchFaceService.Engine.html documentation hidden in a zip file --
thanks to Nick Read for pointing this out.  So instead of relying on
onTimeTick(), install a timer that ticks every 5 minutes and drive
everything off that; in ambient mode too.  This also fixes a FIXME about
not redrawing every minute but only every 5.

Also, the documentation of onVisibilityChanged() makes it very clear
that one has to call super.onVisibilityChanged() as the first thing, so
do this.  This seems to fix the issue of setWatchFaceStyle() not having
an effect immediately on watch face launch.